+++ eras/allegra/impl/test/Test/Cardano/Ledger/Allegra/BinarySpec.hs
@@ -5,7 +5,11 @@ module Test.Cardano.Ledger.Allegra.BinarySpec (spec) where
 import Cardano.Ledger.Allegra
 import Test.Cardano.Ledger.Allegra.Arbitrary ()
 import Test.Cardano.Ledger.Common
-import Test.Cardano.Ledger.Core.Binary as UpgradeSpec
+import Test.Cardano.Ledger.Core.Binary (specUpgrade)
+import Test.Cardano.Ledger.Shelley.Binary.RoundTrip (roundTripShelleyCommonSpec)
 
 spec :: Spec
-spec = specUpgrade @Allegra True
+spec = do
+  specUpgrade @Allegra True
+  describe "RoundTrip" $ do
+    roundTripShelleyCommonSpec @Allegra